Fix: Add Swagger API

+++ items/adapters/item_adapter.py
@@ -10,11 +10,11 @@ class ItemAdapter:
         return self.item.id
 
     def get_price_without_tax(self):
-        return self.item.price_without_tax
+        return float(self.item.price_without_tax)
 
     def get_tax(self):
-        return self.item.tax
+        return float(self.item.tax)
 
     def calculate_price_with_tax(self):
-        return self.get_price_without_tax() * (Decimal(1) + self.get_tax() / Decimal(100))
+        return self.get_price_without_tax() * (float(1) + self.get_tax() / float(100))
